WIP: Proof of Concept tunless http client
This commit hacks in a quick PoC to demonstrate the ability to
connect to a nebula network as a user without the privileges to
create a tun device.

Instead of being a full mesh client, this implementation allows
a single bundled binary to make requests over the nebula network.
Future work could allow for the nebula daemon to act as a SOCKS5
proxy to enable http(s) requests from any process on the host.